+++ src/Types.fs
@@ -5,8 +5,6 @@ open Lmc.ServiceIdentification
 
 // Simple types
 
-type RawData = RawData of FSharp.Data.JsonValue
-
 type CommandId = CommandId of Guid
 
 [<RequireQualifiedAccess>]
@@ -103,3 +101,136 @@ module DataItem =
     let map f item = item |> mapt item.Type f
 
 type Data<'CommandData> = Data of 'CommandData
+
+//
+// Command DTO
+//
+
+type ReactorDto = {
+    Domain: string
+    Context: string
+    Purpose: string
+    Version: string
+    Zone: string
+    Bucket: string
+}
+
+type RequestorDto = {
+    Domain: string
+    Context: string
+    Purpose: string
+    Version: string
+    Zone: string
+    Bucket: string
+}
+
+type ReplyToDto = {
+    Type: string
+    Identification: string
+}
+
+// Generic command data DTO
+
+type DataItemDto<'Value> = {
+    Value: 'Value
+    Type: string
+}
+
+[<RequireQualifiedAccess>]
+module DataItemDto =
+    let serialize serialize (item: DataItem<_>): DataItemDto<_> =
+        {
+            Value = item.Value |> serialize
+            Type = item.Type
+        }
+
+    let internal serializeScalar item = item |> serialize (fun a -> a :> obj)
+
+//
+// Modules
+//
+
+[<RequireQualifiedAccess>]
+module Reactor =
+    let internal serialize (Reactor boxPattern): ReactorDto =
+        {
+            Domain = boxPattern.Domain |> Domain.value
+            Context = boxPattern.Context |> Context.value
+            Purpose = boxPattern.Purpose |> PurposePattern.value
+            Version = boxPattern.Version |> VersionPattern.value
+            Zone = boxPattern.Zone |> ZonePattern.value
+            Bucket = boxPattern.Bucket |> BucketPattern.value
+        }
+
+[<RequireQualifiedAccess>]
+module Requestor =
+    let internal serialize (Requestor box): RequestorDto =
+        {
+            Domain = box.Domain |> Domain.value
+            Context = box.Context |> Context.value
+            Purpose = box.Purpose |> Purpose.value
+            Version = box.Version |> Version.value
+            Zone = box.Zone |> Zone.value
+            Bucket = box.Bucket |> Bucket.value
+        }
+
+[<RequireQualifiedAccess>]
+module ReplyTo =
+    let internal serialize (replyTo: ReplyTo): ReplyToDto =
+        {
+            Type = replyTo.Type
+            Identification = replyTo.Identification
+        }
+
+[<RequireQualifiedAccess>]
+module Data =
+    let data (Data data) = data
+
+//
+// Raw Data
+//
+
+type RawData = RawData of FSharp.Data.JsonValue
+
+[<RequireQualifiedAccess>]
+module RawData =
+    open FSharp.Data
+
+    let value (RawData value) = value
+
+    let (|Itemi|_|) (index, key) (RawData data) =
+        match data with
+        | JsonValue.Record (values) ->
+            try
+                match values.[index] with
+                | k, v when k = key -> Some (Itemi v)
+                | _ -> None
+            with _ -> None
+        | _ -> None
+
+    let (|Item|_|) key (RawData data) =
+        match data with
+        | JsonValue.Record (values) ->
+            values
+            |> Seq.tryPick (function
+                | (name, value) when name = key -> Item (RawData value) |> Some
+                | _ -> None
+            )
+        | _ -> None
+
+    let (|DataItem|_|) data: DataItem<_> option = maybe {
+        let! v =
+            match data with
+            | Item "value" v -> Some (v |> value)
+            | _ -> None
+
+        let! t =
+            match data with
+            | Item "type" t -> Some (t |> value)
+            | _ -> None
+
+        return {
+            Value = v.AsString()
+            Type = t.AsString()
+        }
+    }