+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Add `Command.parse` function
+- Add `RawData` module with active patterns for easier parsing a data out of a command
 
 ## 2.0.0 - 2020-05-13
 - [**BC**] Change all DTO structures to PascalCase, and let the serialization dealt with a format